Move raidenShutdown to toplevel actions, delete store/actions
