--- raiden/src/store/actions.ts
@@ -1,7 +1,0 @@
-import { createStandardAction } from 'typesafe-actions';
-
-import { ShutdownReason } from '../constants';
-
-export const raidenShutdown = createStandardAction('raidenShutdown')<{
-  reason: ShutdownReason | Error;
-}>();
