+++ raiden/tests/e2e/raiden.spec.ts
@@ -15,7 +15,7 @@ import { request } from 'matrix-js-sdk';
 import { Raiden } from 'raiden/raiden';
 import { ShutdownReason } from 'raiden/constants';
 import { initialState } from 'raiden/state';
-import { raidenShutdown } from 'raiden/store/actions';
+import { raidenShutdown } from 'raiden/actions';
 import { newBlock } from 'raiden/channels/actions';
 import { ChannelState } from 'raiden/channels/state';
 import { Storage } from 'raiden/utils/types';
